% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allergy_flu_tweet_count.R
\name{call_twitter}
\alias{call_twitter}
\title{Gets tweets in the recent 6-9 days pertaining to 'asthma', 'flu' & 'allergies'}
\usage{
call_twitter(s4_object, num_tweets, rtweets_token)
}
\arguments{
\item{s4_object}{the object returned from the get_cities function.}

\item{num_tweets}{the number of tweets to be retrieved.}

\item{rtweets_token}{this is the token for accessing twitter}
}
\value{
function returns a dataframe with number of tweets in the past 6-9 days
}
\description{
Gets tweets in the recent 6-9 days pertaining to 'asthma', 'flu' & 'allergies'
}
