% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_forecasts.r
\name{collect_forecasts}
\alias{collect_forecasts}
\title{Generate a collection of forecasts}
\usage{
collect_forecasts(zip_list, weatherAPIkey)
}
\arguments{
\item{zip_list}{vector containing Zip codes for all desired cities.}

\item{weatherAPIkey}{your openweathermap.org API key}
}
\value{
returns a single data frame
}
\description{
This loop calls get_set for each zip code in the city_list generated 
by get_city_list or a vector of Zip codes created by the user directly.
A column is added to the get_set return to log the corresponding Zip code.
If info on a zip code in the list is not available (DNE - 404) 
it is simply left out and prints a message to notify you that there
is no data available for that Zip code.
}
