% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDataDoc.R
\docType{data}
\name{data_changepoint}
\alias{data_changepoint}
\title{To be used with changepoint}
\usage{
data_changepoint
}
\description{
This is a sample weather dataset to show changepoint function will detect
and display the ouput. To be used as part of detect_changepoint function
}
\keyword{data}
