% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_changepoint.R
\name{detect_changepoint}
\alias{detect_changepoint}
\title{Detect changepoints (in means) in weather patterns using bayesian approach}
\usage{
detect_changepoint(data, chains = 1000, burn_in = 200, plot = FALSE)
}
\arguments{
\item{data}{dataframe containing weather information}

\item{chains}{number of markov chains}

\item{burn_in}{number of observations to exclude ( burn-in period)}

\item{plot}{allows the user to plot the changepoint vs. see descriptive summary}
}
\value{
returns the detected changepoint
}
\description{
Detect changepoints (in means) in weather patterns using bayesian approach
}
