% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_plot.R
\name{display_zipcodes}
\alias{display_zipcodes}
\title{Function to display the locations of the selected citie(s)}
\usage{
display_zipcodes(weather_object)
}
\arguments{
\item{weather_object}{S4 object obtained from get_cities and get_current_weather function}
}
\value{
returns a plot with zipcodes plotted on th map
}
\description{
Function to display the locations of the selected citie(s)
}
