% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_plot.R
\name{g_plot}
\alias{g_plot}
\title{plot function}
\usage{
g_plot(df, typ_w, display_legend = "bottom")
}
\arguments{
\item{df}{data frame containing forecasted weather for selected cities}

\item{typ_w}{weather element to display}

\item{display_legend}{toggle disply of legend}
}
\value{
returns plot of forecast data
}
\description{
plot function
}
