% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cities_radius.r
\name{get_city_list}
\alias{get_city_list}
\title{function will create a list of cities around a K mile radius of the specified location}
\usage{
get_city_list(zip_code = NULL, radius = 5)
}
\arguments{
\item{zip_code}{zipcode for the city in reference}

\item{radius}{radius around the primary city of reference}
}
\value{
returns a S4 object with the primary city and list of cities within the specified radius
}
\description{
function will create a list of cities around a K mile radius of the specified location
}
\examples{
get_city_list(zip_code = "10016", radius = 5)
}
