% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.r
\name{get_set}
\alias{get_set}
\title{Generate forecast data set}
\usage{
get_set(zip_code = zip_code, weatherAPIkey = weather_api)
}
\arguments{
\item{zip_code}{zipcode for the desired city}

\item{weatherAPIkey}{your openweathermap.org API key}
}
\value{
returns a data frame
}
\description{
This function will get the set of forecast data for a single 
specified Zip code using your openweathermap.org API key.
}
\examples{
\donttest{
get_set(zip_code = "94305", weatherAPIkey = "your_api_key")
}
}
