% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Generate forecast data set}
\usage{
get_weather(zip_code, weatherAPIkey = weather_api, radius)
}
\arguments{
\item{zip_code}{zipcode for the desired city}

\item{weatherAPIkey}{your openweathermap.org API key}

\item{radius}{(optional) radius centered on specified zip from which to collect data}
}
\value{
returns a data frame
}
\description{
This function when used without the third argument "radius"
will get the forecast data for a single specified Zip code using
your openweathermap.org API key by calling get_set. However, if you
specify a "radius" it will generate a "city_list" S4 object with the
primary city and those within the radius including the Zip code,
city, state, latitude, longitude, and Distance (from original Zip).
It will then loop through these Zip codes to create a full set of forecasts
for the cities within the specified radius.
}
\details{
*Must run setup_weather_api before running examples.*
}
\examples{
\donttest{
a <- get_weather(19901, weather_api)
b <- get_weather(zip_code = 94305, weatherAPIkey = weather_api, radius = 5)
c <- get_weather(94305, weather_api, 5)
}
}
