% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_plot.R
\name{legend_extract}
\alias{legend_extract}
\title{helper function to extract legend}
\usage{
legend_extract(plot_a)
}
\arguments{
\item{plot_a}{a plot from which extract legend}
}
\value{
returns legend from the plot
}
\description{
helper function to extract legend
}
