% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_changepoint.R
\name{plot_changepoint}
\alias{plot_changepoint}
\title{Function to plot changepoint}
\usage{
plot_changepoint(data, change_loc)
}
\arguments{
\item{data}{data}

\item{change_loc}{changepoints}
}
\value{
returns plots
}
\description{
Function to plot changepoint
}
