% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Function to plot forecast data}
\usage{
plot_forecast(weather_object)
}
\arguments{
\item{weather_object}{S4 object obtained from get_cities and get_current_weather function}
}
\description{
Function to plot forecast data
}
