% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDataDoc.R
\docType{data}
\name{sampleStanford}
\alias{sampleStanford}
\title{A sample of the get_weather output for a single city Palo Alto. (zip = 94305)}
\usage{
sampleStanford
}
\description{
This shows the output you can expect if you do not specify a radius
}
\keyword{data}
