% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDataDoc.R
\docType{data}
\name{sampleStanford10mile}
\alias{sampleStanford10mile}
\title{A sample of the get_weather output with a 10 mile radius specified}
\usage{
sampleStanford10mile
}
\description{
This shows the output you can expect if you do specify a radius.
This is for Palo Alto (zip = 94305) and a 10 mile radius.
}
\keyword{data}
