% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_retrieve_api.r
\name{setup_weather_api}
\alias{setup_weather_api}
\title{Function to store api key}
\usage{
setup_weather_api(api_key)
}
\arguments{
\item{api_key}{your api key}
}
\value{
returns a string containing the saved key
}
\description{
This function takes your api key and saves it for convenience
}
