% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyWeather.R
\name{shinyWeather}
\alias{shinyWeather}
\title{This function runs the shiny app}
\usage{
shinyWeather()
}
\description{
Runs a shiny app for an example gui.
MUST FIRST RUN setup_weather_api() and store your api.
Currently the gui will show you a map with the zip codes within the specified
radius marked for visual reference of where your weather readings are coming 
from. Eventually more data can be worked into the displayed map.
}
