% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allergy_flu_tweet_count.R
\name{summarise_tweets}
\alias{summarise_tweets}
\title{helper function to summarise tweets: count number of tweets by day}
\usage{
summarise_tweets(raw_tweet_df)
}
\arguments{
\item{raw_tweet_df}{output from get_twitter function}
}
\value{
returns count of weather specific tweets for each day
}
\description{
helper function to summarise tweets: count number of tweets by day
}
