% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allergy_flu_tweet_count.R
\name{tweet_token_create}
\alias{tweet_token_create}
\title{Function to setup twitter parameters}
\usage{
tweet_token_create(app_name, consumer_key = NULL,
  consumer_secret = NULL, access_token = NULL, access_secret = NULL)
}
\arguments{
\item{app_name}{name of the app}

\item{consumer_key}{consumer key}

\item{consumer_secret}{consumer secret}

\item{access_token}{access token}

\item{access_secret}{access secret}
}
\value{
returns the saved token
}
\description{
This function takes 4 inputs  and saves it
}
