% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_call.r
\name{weather_call}
\alias{weather_call}
\title{Generate weather API call}
\usage{
weather_call(zip_code, weatherAPIkey)
}
\arguments{
\item{zip_code}{Zip code for the desired city}

\item{weatherAPIkey}{your openweathermap.org API key}
}
\value{
returns a string
}
\description{
This function will create the API call for forecast data from/for
a single Zip code to be used in other functions. You could also 
paste the string into your browser to see the raw return if you 
so desire.
}
\examples{
\donttest{
weatherCall(zip_code = 94305, weatherAPIkey = "your_api_key")
weatherCall(94305, "your_api_key")
}
}
