% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather_S4.R
\docType{class}
\name{weather_class-class}
\alias{weather_class-class}
\alias{weather_class}
\alias{show}
\alias{show,weather_class-method}
\alias{save_cities}
\alias{save_cities,weather_class-method}
\alias{save_primary_city}
\alias{save_primary_city,weather_class-method}
\alias{weather_plot_forecasts}
\alias{weather_plot_forecasts,weather_class-method}
\title{An S4 class which holds weather information for cities}
\usage{
show(object)

\S4method{show}{weather_class}(object)

save_cities(object, df)

\S4method{save_cities}{weather_class}(object, df)

save_primary_city(object, df)

\S4method{save_primary_city}{weather_class}(object, df)

weather_plot_forecasts(object)

\S4method{weather_plot_forecasts}{weather_class}(object)
}
\arguments{
\item{object}{S4 object of weather_class}

\item{df}{list of cities within radius 'r' of primary city; returned from get_cities}
}
\description{
Identifies cities within radius and obtains the weather
}
\section{Slots}{

\describe{
\item{\code{primary_city}}{city being referenced (if provided)}

\item{\code{state}}{state to which the city belongs}

\item{\code{zip}}{zipcode of the city (if provided)}

\item{\code{latitude}}{latitude of the city}

\item{\code{longitude}}{longitude of the city}

\item{\code{radius_ref}}{radius in miles}

\item{\code{cities_within_radius}}{dataframe which saves the information about primary city and cities within the radius}

\item{\code{forecast_data}}{dataframe which holds forecast data for cities}

\item{\code{current_data}}{dataframe which holds the current weather data}

\item{\code{twitter_data}}{dataframe holds the number of tweets for each city by date}
}}

